chore: simplify gRPC implementation
