+++ src/bentoml/client.py
@@ -1,0 +1,19 @@
+"""
+Bento Client.
+=============
+
+See https://docs.bentoml.org/en/latest/guides/client.html for more information.
+
+.. code-block:: python
+
+    import bentoml
+
+    client = bentoml.client.Client.from_url("localhost:3000")
+
+    client.predict(np.array([[5.9, 3, 5.1, 1.8]]))
+"""
+from __future__ import annotations
+
+from ._internal.client import Client
+
+__all__ = ["Client"]
